requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.3
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
redis==5.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.18